Refactor expenses table components and update dropdown functionality
- Introduced `ExpensesTable.vue` for improved structure and data handling of expenses.
- Replaced `TableCardDropdown` with `ExpensesTableDropdown` for better clarity and functionality.
- Updated `dashboard.vue` to utilize the new `ExpensesTable` component, enhancing the layout and interaction.
- Adjusted `TableFooter.vue` to refine styling and maintain consistency across table components.
